Delete pre-refactor code
